tester/requirements.txt
redis[asyncio]
httpx
python-dotenv
openai>=1.0.0
asyncio
typing-extensions
